+++ ydb/library/yql/providers/s3/actors/ya.make
@@ -9,17 +9,17 @@ ADDINCL(
 YQL_LAST_ABI_VERSION()
 
 SRCS(
-    raw_read_actor.cpp
     yql_s3_actors_util.cpp
     yql_s3_applicator_actor.cpp
     yql_s3_sink_factory.cpp
+    yql_s3_raw_read_actor.cpp
     yql_s3_source_factory.cpp
     yql_s3_write_actor.cpp
 )
 
 PEERDIR(
-    contrib/libs/fmt
     contrib/libs/apache/arrow
+    contrib/libs/fmt
     contrib/libs/poco/Util
     ydb/library/actors/http
     library/cpp/protobuf/util
@@ -50,9 +50,9 @@ PEERDIR(
 IF (CLANG AND NOT WITH_VALGRIND)
 
     SRCS(
-        source_queue.cpp
         yql_arrow_column_converters.cpp
         yql_s3_read_actor.cpp
+        yql_s3_source_queue.cpp
     )
 
     PEERDIR(
