+++ ydb/library/yql/providers/s3/actors/yql_s3_read_actor.cpp
@@ -36,12 +36,12 @@
 
 #endif
 
-#include "raw_read_actor.h"
-#include "source_queue.h"
 #include "yql_arrow_column_converters.h"
 #include "yql_s3_actors_util.h"
+#include "yql_s3_raw_read_actor.h"
 #include "yql_s3_read_actor.h"
 #include "yql_s3_source_factory.h"
+#include "yql_s3_source_queue.h"
 
 #include <ydb/core/base/events.h>
 #include <ydb/library/services/services.pb.h>
